pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "china-mirror-skills"
version = "0.1.0"
description = "Claude Code skills for optimizing development tools in China network environment"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "China Mirror Skills Contributors"}
]
keywords = ["claude-code", "mirror", "china", "proxy", "pip", "npm", "docker"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Tools",
    "Topic :: System :: Networking",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "jinja2>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/china-mirror-skills"
Repository = "https://github.com/yourusername/china-mirror-skills"
Issues = "https://github.com/yourusername/china-mirror-skills/issues"

[project.scripts]
china-mirror-check = "scripts.check_mirrors:main"
china-mirror-readme = "scripts.render_readme:main"

[tool.setuptools]
packages = []
